correct user to delete posts

+++ hello.py
@@ -99,29 +99,37 @@ def dashboard():
 
 
 @app.route('/posts/delete/<int:id>')
+@login_required
 def delete_post(id):
 	post_to_delete = Posts.query.get_or_404(id)
+	id = current_user.id
+	if id == post_to_delete.poster.id:
+		try:
+			db.session.delete(post_to_delete)
+			db.session.commit()
 
-	try:
-		db.session.delete(post_to_delete)
-		db.session.commit()
+			# Return a message
+			flash("Blog Post Was Deleted!")
 
+			# Grab all the posts from the database
+			posts = Posts.query.order_by(Posts.date_posted)
+			return render_template("posts.html", posts=posts)
+
+
+		except:
+			# Return an error message
+			flash("Whoops! There was a problem deleting post, try again...")
+
+			# Grab all the posts from the database
+			posts = Posts.query.order_by(Posts.date_posted)
+			return render_template("posts.html", posts=posts)
+	else:
 		# Return a message
-		flash("Blog Post Was Deleted!")
+			flash("You Aren't Authorized To Delete That Post!")
 
-		# Grab all the posts from the database
-		posts = Posts.query.order_by(Posts.date_posted)
-		return render_template("posts.html", posts=posts)
-
-
-	except:
-		# Return an error message
-		flash("Whoops! There was a problem deleting post, try again...")
-
-		# Grab all the posts from the database
-		posts = Posts.query.order_by(Posts.date_posted)
-		return render_template("posts.html", posts=posts)
-
+			# Grab all the posts from the database
+			posts = Posts.query.order_by(Posts.date_posted)
+			return render_template("posts.html", posts=posts)
 
 @app.route('/posts')
 def posts():
@@ -141,7 +149,7 @@ def edit_post(id):
 	form = PostForm()
 	if form.validate_on_submit():
 		post.title = form.title.data
-		post.author = form.author.data
+		#post.author = form.author.data
 		post.slug = form.slug.data
 		post.content = form.content.data
 		# Update Database
@@ -150,7 +158,7 @@ def edit_post(id):
 		flash("Post Has Been Updated!")
 		return redirect(url_for('post', id=post.id))
 	form.title.data = post.title
-	form.author.data = post.author
+	#form.author.data = post.author
 	form.slug.data = post.slug
 	form.content.data = post.content
 	return render_template('edit_post.html', form=form)
@@ -163,11 +171,12 @@ def add_post():
 	form = PostForm()
 
 	if form.validate_on_submit():
-		post = Posts(title=form.title.data, content=form.content.data, author=form.author.data, slug=form.slug.data)
+		poster = current_user.id
+		post = Posts(title=form.title.data, content=form.content.data, poster_id=poster, slug=form.slug.data)
 		# Clear The Form
 		form.title.data = ''
 		form.content.data = ''
-		form.author.data = ''
+		#form.author.data = ''
 		form.slug.data = ''
 
 		# Add post data to database
@@ -375,10 +384,11 @@ class Posts(db.Model):
 	id = db.Column(db.Integer, primary_key=True)
 	title = db.Column(db.String(255))
 	content = db.Column(db.Text)
-	author = db.Column(db.String(255))
+	#author = db.Column(db.String(255))
 	date_posted = db.Column(db.DateTime, default=datetime.utcnow)
 	slug = db.Column(db.String(255))
-
+	# Foreign Key To Link Users (refer to primary key of the user)
+	poster_id = db.Column(db.Integer, db.ForeignKey('users.id'))
 
 # Create Model
 class Users(db.Model, UserMixin):
@@ -390,6 +400,9 @@ class Users(db.Model, UserMixin):
 	date_added = db.Column(db.DateTime, default=datetime.utcnow)
 	# Do some password stuff!
 	password_hash = db.Column(db.String(128))
+	# User Can Have Many Posts 
+	posts = db.relationship('Posts', backref='poster')
+
 
 	@property
 	def password(self):
